Resize Observe & Understand RabbitMQ slides to 720p
They were originally designed for 4:3 (1024x768), but the actual
resolution turned out to be 1080p. While it couldn't addressed on the
spot, and re-recording all videos & screenshots now is not worth the
effort, the best fix that is reasonably easy to implement is to contain
everything height-wise, and put a dark background on the slides that
have dark screenshots or videos.

Chose 720p over 1080p so that slides would look bigger on 1080p
resolutions. Big is good for slides.